saving...
have sidenav clicks transferred to summary view page
